block infinity for enqueue operation
Although the enqueue has to be as fast as possible (it is called inside
the publish dispatch loop) we have to know if the message got
dispatched, otherwise the dispatch-loop has to persist the message in
case of a dispatch error for QoS1 & 2. A Timeout error won't tell us if
the message hasn't been dispatched (late gen_fsm 'ok' replies can occur
in times of high load). Infinity seems the best option we currently
have. Therefore we had to remove the blocking behaviour of possible
remote msg store 'deref' calls.