use of new plumtree_metadate implementation
The new plumtree_metadata implementation doesn't use an ETS frontend,
this heavily reduces memory consumption, as all the metadata
(subscriptions and retained messages) was stored in ETS as well as in
LevelDB. We ditched ETS in favor of using LevelDB directly. As a
consequence the iterator interface changed a bit, which forced us to
change the way we deal with delivering retained messages. Since this was
an open issue anyways (performance) we now serve the retained messages
through an ETS table directly (which is filled by using a
plumtree_metadata subscription).

Good enough reason for such a big commit.