Make random tests reproducible via a seed
This introduces a couple of helper functions:

     vmq_test_utils:get_suite_rand_seed/0

Generates a seed from `os:timestamp()` for the entire suite if one is
not set via `{seed, Seed}.` in a config file passed to CT. This function
should be called from the `init_per_suite/1` function and the seed
returned should be added to the CT `Config` list. The seed will be
printed to console.

     vmq_test_utils:seed_rand/1

Is for use in the individual test cases to seed the rng.

     vmq_test_utils:rand_bytes/1

This replaces the use of `crypto:strong_rand_bytes` in tests so random
bytes will be deterministically drawn according to the seed.