Work around cs true cleanup race
When connecting and publishing sequentially to several nodes with the
same client id and clean session true, we sometimes trigger a queue
cleanup which is due to the eventually consistent data-store.

Connecting and publishing to random nodes is not needed to test what the
test was built for, so for now we just publish to a single node so we
don't trigger the cleanup race.