Make retain suite run faster
Make the testcases independent by using distinct topics and
client-ids which removes the necessity of tearing down and restarting
vmq_server between cases. Further since, all tests cases use
independent sessions and topics they can run in parallel. This reduces
the runtime of the suite to about 11 seconds where it originally ran
in 23 seconds.

Further the specific auth hooks matching on topics and client-ids has
been simplified to just accept everything as the purpose of this test
is not to test the hook system.