Add compat feature for module plugins
Fixed a bug in `partition_hooks` so hooks with compat information are
treated as different from hooks without compat information.

Fixed a bug in the check if a plugin already has a specific hook by
also taking the compat information into account. Otherwise the same
hook with and without compat information would look the same and one
would be rejected.