Fix cluster leave low timeout issue
This removes the race-condition that would occur if the spawned
process for the `LeaveFun` had exited before the monitor call was
madeas this would have returned the atom `noproc` to the cli which
couldn't be converted to a string, causing a badarg exception.

Further, the logic is changed so `Iterations > 0` which ensures the
logic checking if queue migration is complete is run at least
once. Before it would be skipped if the timeout was less than 5
seconds.