Remove cuttlefish warning about message_size_limit
The setting `message_size_limit` has been deprecated for long but
still when one runs `vernemq config generate` or starts the broker
without a configuration file, it warns about this setting even when is
not set. This change makes it warn only if the setting has been
changed to a non-default value.