Merge master into mqtt5 preview (#745)
* Improve cluster leave and cluster readiness probing in general (#731)

`is_reachable/1` is replaced by `status/1` which can return `up`,
`down`, or `init` (instead of the old boolean). A cluster node process
with the status `init` has been started BUT wasn't able to establish a
connection. This case is different than a cluster node process that
looses an established connection. The former is a pretty normal case
when setting up multiple nodes in parallel, while the latter indicates a
network partition and therefore ends in a `down` status. As a
consequence a `init` state wont harm the cluster availability.

Additionally the cluster readiness probe counts the number of detected
and resolved netsplit situations. Those counters are also exposed as
metrics `netsplit_detected` and `netsplit_resolved`.

* Fix readyness probe in a healthy cluster that has previously resolved a netsplit (#742)

* Prepare MQTTv5 message and cluster changes

- Make it possible to extend the cluster internal message format when
  adding new features while ensuring back- and forwards compatibility.
- Refactor the leveldb message store code to support versioned indexes
  as well as messages.
- Refactor the retained message store to support versions.
- Add support for subscriber data with subscription options.