Fix QoS2 retry for reconnecting clients, and retry mechanism improvement
- Store the *parsed* PUBREL inside the waiting acks instead of the
serialized binary.  This enables to reuse the same functionality for
handling retried PUBRELs, without caring if it is retried because of a
timeout or because of a reconnect. Cleanup places which relied on the
serialized binary.
- Fixing above enabled the possibility to implement a small performance
improvement where we could get rid of using `length()` on the outgoing
batch of frames for incrementing the `mqtt_publishes_sent` counter.
- Improved `vmq_publish_SUITE.erl` to test that a client responds to
PINGREQs before closing the Socket. This validates that 1st. the client
hasn't crashed in the meantime (e.g. because not being able to handle
the PUBCOMP, see bug #762), and 2nd that the last received frame on this
socket is the PINGRESP.
- The improvement above discovered an edgecase with a PUBREL frame that
has been retried too fast, indicating a bug in the retry mechanism. This
has been fixed with tagging the message IDs inside the retry queue. As a
result the retry mechanism can differentiate between retrying a PUBLISH
and a PUBREL. Without this fix a retried, already acked PUBLISH, would
have retried an unacked PUBREL right away.
- Removing several function clauses that were forgotten when fixing #750