Add optional vmq_swc plugin
vmq_swc is an alternative to vmq_plumtree and provides a metadata
storage backed by RocksDB and uses Server Wide Clocks (SWC) for
optimistic replication. SWC is a novel distributed algorithm that
enables VerneMQ to better scale subscription and retained message
data and is a hot candidate to replace the Plumtree Metadata
found in the current VerneMQ version. The aim is to heavily
improve startup times with 1M+ subscribers, CPU and memory profile,
and communication overhead (by using batching wherever possible),
moreover it solves the 'hard' problem of distributed deletes in an
eventual consistent world.

Currently you must build VerneMQ with `make swc` to generate a
release that has vmq_swc bundled and configured. In such a release
the vmq_plumtree wont be available.