Fix `vmq-admin session show` crash when using `--limit=X`
When using `--limit=X` we would ask for *any* fields from any session
if no fields were specified. Since some sessions are offline and some
are online, a set of rows with differing number of colums would be
returned (with and without the `session_pid`) which would cause a
crash in `clique` when preparing the output.

This fix ensures we can never pass `*` as the selection criterion but
will use a default list of fields if none were specified.

This would have the added benefit from being more efficient that less
data is return by default when `--limit=X` is used.