Make listener configuration independent of the config change hook
To simplify a bit this change removes the ability to reconfigure the
listeners when the `on_config_change_hook`. As we earlier didn't have
a well defined point in the supervision hierarcy to start the
listeners we needed to implement the hook - but now we can just do the
initialization directly.