Make conversion of modifiers returned from Lua context aware
We need to have a context when converting modifiers returned from Lua
in order to handle MQTT 5.0 properties correctly. In particular the
`vmq_diversity_utils:convert/1` function cannot be used to convert the
user properties as it would convert the user property keys to atoms.

So now all conversion has been moved into the `vmq_diversity_plugin`
where we can hint the conversion which hook is being invoked and
handle the special cases directly.