Fix errors logs from health endpoint
The endpoint was a `cowboy_rest` endpoint it and the cowboy rest code
doesn't like it when `cowboy_req:reply` is used to send the
response. This is likely something which changed since the cowboy 1.x
series which we missed when upgrading to cowboy 2.x.

The endpoint was changed to be a plain http handler where we have more
control over the replies and error codes.