Fix shared subscription bug that use the MQTTv5 Subscription Identifier property
The bug was triggered if a QoS > 0 message gets routed to a shared
subscription which uses the MQTTv5 Subscription Identifier propery. In
such a case the QoS wasn't correctly set for the message, resulting in a
corrupt message on disk. However one only experiences the issue if such
a 'corrupt' message was read from the message store as part of the
decompress step (maybe_deref/2).