Loosen MongoDB reads
This fixes the issue with r_mode not working in the latest version
of MongoDB. The r_mode option is deprecated in the underlying
MongoDB driver. We must not set a read preference instead.

This PR uses the primaryPreferred read preference by default.
However, the value isn't really meaningful, since VerneMQ
Diversity's pooling mechanism does not monitor a topology
of primary and secondary nodes in a MongoDB replica set.
This default read preference is a reasonable choice for
folks using VMQ Diversity to auth against MongoDB.