Centralize JSON encoding/decoding
This PR abstracts JSON encoding/decoding in a common module, despite of
what library is used behind the scenes. This allows for easier refactoring
as well as switching the JSON library to a faster one with minimal effort.

With upcoming JSON logging that I'm working on it would probably make a difference
to have a library like jsone or thoas that are faster and more memory efficient than JSX.