vmq_tracer_cli: Read on STDIN instead of waiting for group leader exit
In order to make tracing easier to use from non-interactive programs, we
read on STDIN in a blocking fashion to learn when the pipe is closed.

Simply monitoring the group leader process is not enough when the tracer
is invoked by a non-interactive program as the STDIN/STDOUT pipes might
be closed without SIGINT being transmitted to the erlang process.

Reading on STDIN works for both interactive and non-interactive use
cases.